Extract FeatureDocumentReference from GroundFirestore
